+++ src/screen_title.h
@@ -7,17 +7,17 @@ class TitleScreen final : public Screen
 public:
 	TitleScreen();
 	void init() override;
-	void update(float delta) override;
-	void draw(float delta) override;
+	void update(f32 delta) override;
+	void draw(f32 delta) override;
 	void unload() override;
 	const char *getName() const override { return "TitleScreen"; }
 
 private:
-	static constexpr float BANNER_STARTPOS = -150.0f;
-	static constexpr float BANNER_ENDPOS = 32.0f;
+	static constexpr i32 BANNER_STARTPOS = -150.0f;
+	static constexpr i32 BANNER_ENDPOS   = 32.0f;
 
 	GameMode mode_{ GameMode::LOREM };
-	int hiscore_;
-	float bannerPosition_ = BANNER_STARTPOS;
-	float animTime_ = 0.0f;
+	u64 hiscore_;
+	f32 bannerPosition_ = BANNER_STARTPOS;
+	f32 animTime_       = 0.0f;
 };
