+++ src/viewport.h
@@ -1,4 +1,6 @@
 #pragma once
+#include "utils.h"
+
 #include <raylib.h>
 
 class Viewport
@@ -11,14 +13,14 @@ public:
 	void endDrawing() const;
 	void drawRenderTexture() const;
 
-	static constexpr int GAME_WIDTH = 512;
-	static constexpr int GAME_HEIGHT = 448;
-	static constexpr int WINDOW_WIDTH = 1024;
-	static constexpr int WINDOW_HEIGHT = 896;
+	static constexpr i16 GAME_WIDTH    = 512;
+	static constexpr i16 GAME_HEIGHT   = 448;
+	static constexpr i16 WINDOW_WIDTH  = 1024;
+	static constexpr i16 WINDOW_HEIGHT = 896;
 
 private:
-	float virtualRatio_ = 0.0f;
+	f32 virtualRatio_              = 0.0f;
 	RenderTexture2D renderTexture_ = {};
-	Rectangle sourceRect = {};
-	Rectangle destRect = {};
+	Rectangle sourceRect           = {};
+	Rectangle destRect             = {};
 };
